big refactoring on functions params in all parts of the code. Added flexibility on new solvers additions. Removed duplication in exports from all solvers.

+++ R/cleanup_carnival.R
@@ -2,39 +2,38 @@
 ##
 ## Enio Gjerga, 2020
 
-#TODO update 
 cleanupCARNIVAL <- function(carnivalOptions){
-  
-  message("Cleaning intermediate files")
-  
-  lpFile <- carnivalOptions$filenames$lpFilename 
-  resultFile <- carnivalOptions$filenames$resultFile
-  keepLpFiles <- carnivalOptions$keepLPFiles
-  
-  if(!keepLpFiles & file.exists(lpFile)){
-    file.remove(lpFile)
-  }
-  
-  if(file.exists(resultFile)){
-    file.remove(resultFile)
-  }
-  
-  if (carnivalOptions$solver == supportedSolvers$cplex) {
-    workdir <- carnivalOptions$workdir
-    cplexCommandFile <- carnivalOptions$filenames$cplexCommandFile
+
+  if (carnivalOptions$cleanTmpFiles) {
+    message("Cleaning intermediate files")
+    lpFile <- carnivalOptions$filenames$lpFilename 
+    resultFile <- carnivalOptions$filenames$resultFile
+    keepLpFiles <- carnivalOptions$keepLPFiles
     
-    if(file.exists("cplex.log")){
-      file.remove("cplex.log")
+    if(!keepLpFiles & file.exists(lpFile)){
+      file.remove(lpFile)
     }
     
-    if(file.exists(cplexCommandFile)){
-      file.remove(cplexCommandFile)
+    if(file.exists(resultFile)){
+      file.remove(resultFile)
     }
     
-    cloneFiles <- list.files(path = workdir, pattern = "$clone")
-    file.remove(cloneFiles)
+    if (carnivalOptions$solver == supportedSolvers$cplex) {
+      workdir <- carnivalOptions$workdir
+      cplexCommandFile <- carnivalOptions$filenames$cplexCommandFile
+      
+      if(file.exists("cplex.log")){
+        file.remove("cplex.log")
+      }
+      
+      if(file.exists(cplexCommandFile)){
+        file.remove(cplexCommandFile)
+      }
+      
+      cloneFiles <- list.files(path = workdir, pattern = "$clone")
+      file.remove(cloneFiles)
+    }
+    
+    message("Done: cleaning")
   }
-  
-  message("Done: cleaning")
-  
 }
